requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-rest-auth==2.2.4
Django==4.0.5
django-cors-headers==3.13.0
django-filter==22.1
django-simpleui==2022.6.131
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
